+++ packages/firestore/lite/src/api/database.ts
@@ -20,14 +20,20 @@ import { _FirebaseService, FirebaseApp } from '@firebase/app-types-exp';
 import { Provider } from '@firebase/component';
 
 import { Code, FirestoreError } from '../../../src/util/error';
-import { DatabaseId } from '../../../src/core/database_info';
+import { DatabaseId, DatabaseInfo } from '../../../src/core/database_info';
 import { FirebaseAuthInternalName } from '@firebase/auth-interop-types';
 import {
   CredentialsProvider,
+  CredentialsSettings,
   FirebaseCredentialsProvider
 } from '../../../src/api/credentials';
 import { removeComponents } from './components';
-import { FirestoreSettings } from '../../../src/api/database';
+import {
+  LRU_COLLECTION_DISABLED,
+  LRU_DEFAULT_CACHE_SIZE_BYTES,
+  LRU_MINIMUM_CACHE_SIZE_BYTES
+} from '../../../src/local/lru_garbage_collector';
+import { validateIsNotUsedTogether } from '../../../src/util/input_validation';
 
 declare module '@firebase/component' {
   interface NameServiceMapping {
@@ -35,10 +41,107 @@ declare module '@firebase/component' {
   }
 }
 
+// settings() defaults:
+const DEFAULT_HOST = 'firestore.googleapis.com';
+const DEFAULT_SSL = true;
+
 export interface Settings {
   host?: string;
   ssl?: boolean;
   ignoreUndefinedProperties?: boolean;
+  cacheSizeBytes?: number;
+  experimentalForceLongPolling?: boolean;
+  experimentalAutoDetectLongPolling?: boolean;
+}
+
+/** Undocumented, private additional settings not exposed in our public API. */
+interface PrivateSettings extends Settings {
+  // Can be a google-auth-library or gapi client.
+  credentials?: CredentialsSettings;
+}
+
+/**
+ * A concrete type describing all the values that can be applied via a
+ * user-supplied firestore.Settings object. This is a separate type so that
+ * defaults can be supplied and the value can be checked for equality.
+ */
+export class FirestoreSettings {
+  /** The hostname to connect to. */
+  readonly host: string;
+
+  /** Whether to use SSL when connecting. */
+  readonly ssl: boolean;
+
+  readonly cacheSizeBytes: number;
+
+  readonly experimentalForceLongPolling: boolean;
+
+  readonly experimentalAutoDetectLongPolling: boolean;
+
+  readonly ignoreUndefinedProperties: boolean;
+
+  // Can be a google-auth-library or gapi client.
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  credentials?: any;
+
+  constructor(settings: PrivateSettings) {
+    if (settings.host === undefined) {
+      if (settings.ssl !== undefined) {
+        throw new FirestoreError(
+          Code.INVALID_ARGUMENT,
+          "Can't provide ssl option if host option is not set"
+        );
+      }
+      this.host = DEFAULT_HOST;
+      this.ssl = DEFAULT_SSL;
+    } else {
+      this.host = settings.host;
+      this.ssl = settings.ssl ?? DEFAULT_SSL;
+    }
+
+    this.credentials = settings.credentials;
+    this.ignoreUndefinedProperties = !!settings.ignoreUndefinedProperties;
+
+    if (settings.cacheSizeBytes === undefined) {
+      this.cacheSizeBytes = LRU_DEFAULT_CACHE_SIZE_BYTES;
+    } else {
+      if (
+        settings.cacheSizeBytes !== LRU_COLLECTION_DISABLED &&
+        settings.cacheSizeBytes < LRU_MINIMUM_CACHE_SIZE_BYTES
+      ) {
+        throw new FirestoreError(
+          Code.INVALID_ARGUMENT,
+          `cacheSizeBytes must be at least ${LRU_MINIMUM_CACHE_SIZE_BYTES}`
+        );
+      } else {
+        this.cacheSizeBytes = settings.cacheSizeBytes;
+      }
+    }
+
+    this.experimentalForceLongPolling = !!settings.experimentalForceLongPolling;
+    this.experimentalAutoDetectLongPolling = !!settings.experimentalAutoDetectLongPolling;
+
+    validateIsNotUsedTogether(
+      'experimentalForceLongPolling',
+      settings.experimentalForceLongPolling,
+      'experimentalAutoDetectLongPolling',
+      settings.experimentalAutoDetectLongPolling
+    );
+  }
+
+  isEqual(other: FirestoreSettings): boolean {
+    return (
+      this.host === other.host &&
+      this.ssl === other.ssl &&
+      this.credentials === other.credentials &&
+      this.cacheSizeBytes === other.cacheSizeBytes &&
+      this.experimentalForceLongPolling ===
+        other.experimentalForceLongPolling &&
+      this.experimentalAutoDetectLongPolling ===
+        other.experimentalAutoDetectLongPolling &&
+      this.ignoreUndefinedProperties === other.ignoreUndefinedProperties
+    );
+  }
 }
 
 /**
@@ -193,4 +296,19 @@ export function getFirestore(app: FirebaseApp): FirebaseFirestore {
 export function terminate(firestore: FirebaseFirestore): Promise<void> {
   _removeServiceInstance(firestore.app, 'firestore/lite');
   return firestore._delete();
+}
+
+export function makeDatabaseInfo(
+  databaseId: DatabaseId,
+  persistenceKey: string,
+  settings: FirestoreSettings
+): DatabaseInfo {
+  return new DatabaseInfo(
+    databaseId,
+    persistenceKey,
+    settings.host,
+    settings.ssl,
+    settings.experimentalForceLongPolling,
+    settings.experimentalAutoDetectLongPolling
+  );
 }
